Start to flush out the documentation
Now that we have some docstrings setup this commit starts the process of
flushing out the library documentation. This adds a new doc page on
using the API, which will go into more depth on how to use all the
aspects of the pyopnsense API. Right now it just gives a high level
explanation similar to what's in the README, but with a little extra
detail.