Add official support for Python 3.8 and Python 3.9
This commit adds support for running on Python 3.8 and 3.9. Nothing
actually prevented it from running before but the trove classifiers in
the package metadata and CI jobs were missing for these newer python
versions. This commit addresses this by adding these to the package
metadata and adding CI jobs for Python 3.8 and 3.9.