Update supported python versions and CI config
This commit updates the supported list of supported Python versions to
only include currently supported upstream versions of cPython. From this
commit onward only Python 3.7 will be supported. This also enables us to
drop the six dependency as it was only used to maintain Python 2.7
compat. Additionally the Ci configuration is updated to only run tests
for the supported versions of Python.