Pivot to github actions for CI
The repo was previously using a combination of travis and azure-pipelines (for windows) for CI. Since those were introduced travis stopped being a viable platform for use with open source projects and the configuration for azure went stale. This commit switches to use github actions instead as it is maintained and offers generous quotas for open source projects.