Add black to CI/dev workflow (#22)
This commit adds [black](https://github.com/psf/black) to the
development workflow for pyopnsense. Using an autoformatter will
simplify development because we no longer have to worry about formatting
code with a consistent style used in the project because black will just
make the changes necessary automatically.

Black has been added to all the CI configurations and a full run has
been made on the repository updating the formatting.